ServerSentEvents/submissions/MOSQUITO/resolve-url-argument.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8" />
	<title>When the EventSource() constructor is invoked, the UA must run these steps: Resolve the URL specified in the first argument:</title>
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>When the EventSource() constructor is invoked, the UA must run these steps: Resolve the URL specified in the first argument:</h1>

	<div id="log"></div>
	<div id="test" style="display:block"></div>
  <script>
    test(function() {assert_throws("SYNTAX_ERR", function(){new EventSource('http://invalid url/');})}, "If the URL specified (http://invalid url/) in the first argument is false, then throw a 'SYNTAX_ERR' exception.");
    test(function() {assert_throws("SECURITY_ERR", function(){new EventSource('http://googlecodesamples.com/html5/sse/sse.php');})}, "If the URL specified (http://googlecodesamples.com/html5/sse/sse.php) in the first argument is not the same source, then throw a 'SECURITY_ERR' exception.");
    test(function() {new EventSource('http://html5.mosquito-fp7.eu/messages/t0001');}, "If the URL specified (http://html5.mosquito-fp7.eu/messages/t0001) in the first argument is true, then not throw a 'SECURITY_ERR' exception.");
	/*
	var s = new EventSource('http://googlecodesamples.com/html5/sse/sse.php');
	s.onmessage = function(e){
		console.log(e.data);
	};
	//*/
  </script>
  </div>
</body>
</html>
